调整超时时间设置位置，以及回调的超时时间。

+++ joyrpc-core/src/main/java/io/joyrpc/protocol/message/RequestMessage.java
@@ -289,7 +289,7 @@ public class RequestMessage<T> extends BaseMessage<T> implements Request {
         this.transport = transport;
     }
 
-    public long getTimeout() {
+    public int getTimeout() {
         return timeout;
     }
